more client work to mesophyll

+++ src/mesophyll/message.rs
@@ -26,6 +26,6 @@ pub enum ClientMessage {
     },
 
     Heartbeat { 
-        vm_count: u64 
+        vm_count: usize 
     },
 }
